// src/kafka/topics/sendEmail/sendEmail.enum.ts
export enum SendEmails {
  sendEmail = 'send-email',
  sendEmailKYC = 'send-email-kyc',
  sendEmailDynamic = 'send-email-dynamic',
  sendEmailViaHtml = 'send-email-html',
  sendLeadEmailViaHtml = 'send-lead-email-html',
  updateDraftEmailViaHtml = 'update-draft-email-html',
  sendCommunication = 'send-communication',
  addEmailTemplate = 'add-email-template',
  updateEmailTemplate = 'update-email-template',
  deleteEmailTemplate = 'delete-email-template',
  updateCommunication = 'update-communication',
  sendWeclomeMails = 'send-welcome-mails',
  sendAgreementEmail = 'send-agreement-mails',
  getAllCommunication = 'get-all-communications',
  getLeadCommunication = 'get-lead-communications',
  getAllTemplates = 'get-all-templates',
  getAllAgreements = 'get-all-user-agreements',
  markMessageAsRead = 'mark-read',
  getOneMessage = 'get-one-message',
  getOneTemplate = 'get-one-template',
  getEmailInbox = 'get-email-inbox',
  getEmailInboxByLead = 'get-email-inbox-lead',
  getEmailInboxById = 'get-email-inbox-detail',
  sendMeetingMails = 'send-meeting-mails',
  getLatestLead = 'get-latest-communications',
  sendOtp = 'send-otp',
  testSmtp = 'get-test-smtp'
}